dbeaver/pro#10008: refactor profile logic detach to manager
